Move io_support to vtk_interface
